@@ -3,13 +3,16 @@
   "version": "1.0.0",
   "description": "☁️ open source continuous deployment service",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "test": "tap test/index.js test/lib",
-    "coverage": "tap test/index.js test/lib --coverage --coverage-report=lcov",
+    "test": "jest",
+    "coverage": "jest --collect-coverage",
     "lint": "eslint .",
-    "start": "./bin/deploy serve",
+    "start": "./bin/deploy.js serve",
     "pack": "pkg bin/deploy.js -c package.json -o packed/deploy",
-    "generate-docs": "tryitout && jsdoc -c jsdoc.json"
+    "generate-docs": "node generate-docs.js",
+    "prepare": "husky install",
+    "release": "release-it"
   },
   "pkg": {
     "scripts": [
@@ -20,10 +23,10 @@
       "lib/static/*"
     ],
     "targets": [
-      "node8-alpine-x64",
-      "node8-linux-x64",
-      "node8-macos-x64",
-      "node8-win-x64"
+      "node16-alpine-x64",
+      "node16-linux-x64",
+      "node16-macos-x64",
+      "node16-win-x64"
     ]
   },
   "repository": {
@@ -34,33 +37,40 @@
     "deploy": "./bin/deploy.js"
   },
   "engines": {
-    "node": ">= 8"
+    "node": ">= 16"
   },
-  "homepage": "https://www.gabrielcsapo.com/deploy.sh",
+  "homepage": "https://gabrielcsapo.github.io/deploy.sh",
   "dependencies": {
-    "body-parser": "^1.17.2",
-    "commander": "^2.11.0",
-    "dockerode": "^2.5.0",
-    "express": "^4.15.4",
-    "formidable": "^1.1.1",
-    "inquirer": "^3.2.1",
-    "moment": "^2.18.1",
-    "mongoose": "^4.11.6",
-    "opn": "^5.1.0",
-    "ora": "^1.3.0",
-    "request": "^2.81.0",
-    "tar": "^4.3.0",
-    "turtler": "^1.0.2",
-    "update-notifier": "^2.2.0",
-    "woof": "^0.2.1"
+    "axios": "^0.26.1",
+    "body-parser": "^1.19.2",
+    "commander": "^9.0.0",
+    "dockerode": "^3.3.1",
+    "express": "^4.17.3",
+    "form-data": "^4.0.0",
+    "formidable": "^2.0.1",
+    "inquirer": "^8.2.1",
+    "moment": "^2.29.1",
+    "mongoose": "^6.2.6",
+    "opn": "^6.0.0",
+    "ora": "^6.1.0",
+    "tar": "^6.1.11",
+    "turtler": "^1.1.1",
+    "update-notifier": "^5.1.0",
+    "woof": "^0.4.1"
   },
   "devDependencies": {
-    "eslint": "^4.4.1",
-    "jsdoc": "^3.5.4",
-    "minami": "^1.2.3",
-    "pkg": "^4.3.0",
-    "tap": "^10.7.1",
-    "tape": "^4.8.0",
-    "tryitout": "^2.0.7"
+    "eslint": "^8.11.0",
+    "eslint-plugin-jest": "^26.1.1",
+    "husky": "^7.0.0",
+    "jest": "^27.5.1",
+    "jsdoc": "^3.6.10",
+    "jsdoc-to-markdown": "^7.1.1",
+    "pkg": "^5.5.2",
+    "prettier": "^2.5.1",
+    "pretty-quick": "^3.1.3",
+    "release-it": "^14.12.5"
+  },
+  "volta": {
+    "node": "16.14.0"
   }
 }
