@@ -1,31 +1,36 @@
 #!/usr/bin/env node
 
-const fs = require('fs');
-const path = require('path');
+import fs from "fs";
+import path from "path";
+import ora from "ora";
 
-module.exports = async function(cli, spinner) {
-  spinner.text = 'Creating application bundle';
+export default async function (cli) {
+  const spinner = ora().start();
+
+  spinner.text = "Creating application bundle";
 
   await cli.createBundle(process.cwd());
 
-  spinner.text = 'Getting deploy keys';
+  spinner.text = "Getting deploy keys";
 
   const { token, username } = await cli.getCredentials();
 
-  spinner.text = 'Uploading application bundle';
+  spinner.text = "Uploading application bundle";
 
-  const bundle = fs.createReadStream(path.resolve(process.cwd(), 'bundle.tgz'));
+  const bundle = fs.readFileSync(path.resolve(process.cwd(), "bundle.tgz"));
 
   const { deployment } = await cli.uploadBundle({
     name: path.basename(process.cwd()),
     bundle,
     token,
-    username
+    username,
   });
 
-  spinner.text = 'Cleaning up local files';
+  spinner.text = "Cleaning up local files";
 
   await cli.removeBundle(process.cwd());
 
-  spinner.succeed(`Upload successfully, http://${deployment.subdomain}.localhost:5000`);
-};
+  spinner.succeed(
+    `Upload successfully, http://${deployment.subdomain}.localhost:5000`
+  );
+}
