@@ -1,12 +1,16 @@
 #!/usr/bin/env node
 
-module.exports = async function(cli, spinner) {
+import ora from "ora";
+
+export default async function (cli) {
+  const spinner = ora().start();
+
   const { token, username } = await cli.getCredentials();
 
   spinner.text = `Logging out of current session for ${username}`;
 
   await cli.logout({ token, username });
-  await cli.cacheCredentials({ username: '', token: '' });
+  await cli.cacheCredentials({ username: "", token: "" });
 
-  spinner.succeed(`Logged out of session for ${username} successfully`);
-};
+  spinner.succeed(`Logged out as ${username}`);
+}
