@@ -1,19 +1,26 @@
 #!/usr/bin/env node
 
-module.exports = async function(cli, spinner) {
-  spinner.text = `Getting logs for ${cli.application}`;
+import ora from "ora";
+
+export default async function (cli) {
+  const spinner = ora().start();
+
+  spinner.text = `Getting logs for application: "${cli.application}"`;
 
   const { token, username } = await cli.getCredentials();
-  const { logs } = await cli.getLogs({ token, username, name: cli.application });
+  const { logs } = await cli.getLogs({
+    token,
+    username,
+    name: cli.application,
+  });
 
-  spinner.stop();
-
-  if(logs) {
-    console.log('' + // eslint-disable-line
-    `==========
-      ${logs.join('').trim()}
-    ==========`);
+  if (logs) {
+    spinner.info(`
+==========
+${logs.join("").trim()}
+==========
+    `);
   } else {
-    console.log('No logs available 🙈'); // eslint-disable-line
+    spinner.info("No logs available 🙈"); // eslint-disable-line
   }
-};
+}
