@@ -1,13 +1,20 @@
 #!/usr/bin/env node
 
-const Url = require('url');
-const opn = require('opn');
+import Url from "url";
+import opn from "opn";
+import ora from "ora";
 
-module.exports = async function(cli, spinner) {
-  spinner.text = 'Opening up url to deployment instance';
+export default async function (cli) {
+  const spinner = ora().start();
+
+  spinner.text = "Opening up url to deployment instance";
 
   const { token, username } = await cli.getCredentials();
-  const { deployment } = await cli.getDeployments({ token, username, name: cli.application });
+  const { deployment } = await cli.getDeployments({
+    token,
+    username,
+    name: cli.application,
+  });
 
   const config = Url.parse(cli.url);
   config.host = `${deployment.subdomain}.${config.host}`;
@@ -16,4 +23,4 @@ module.exports = async function(cli, spinner) {
   spinner.text = `Opening deployment at ${url}`;
   spinner.stopAndPersist();
   opn(url, { wait: false });
-};
+}
