@@ -1,8 +1,8 @@
-const fs = require('fs');
-const path = require('path');
-const { promisify } = require('util');
+import fs from "fs";
+import path from "path";
+import { promisify } from "util";
 
-const { contains } = require('./helpers/util');
+import { contains } from "./helpers/util.js";
 
 const readFile = promisify(fs.readFile);
 
@@ -15,19 +15,21 @@ const readFile = promisify(fs.readFile);
 async function classifier(directory) {
   const files = fs.readdirSync(directory);
 
-  if(contains(files, ['Dockerfile'])) {
-    let build = (await readFile(path.resolve(directory, 'Dockerfile'))).toString('utf8');
+  if (contains(files, ["Dockerfile"])) {
+    let build = (
+      await readFile(path.resolve(directory, "Dockerfile"))
+    ).toString("utf8");
     build = build.replace(/\$\{directory\}/, directory);
 
     return {
-      type: 'docker',
-      build
+      type: "docker",
+      build,
     };
   }
 
-  if(contains(files, ['package.json'])) {
+  if (contains(files, ["package.json"])) {
     return {
-      type: 'node',
+      type: "node",
       build: `
         FROM mhart/alpine-node:8
         WORKDIR ${directory}
@@ -36,26 +38,26 @@ async function classifier(directory) {
         RUN npm install
 
         CMD ["npm", "start"]
-      `
+      `,
     };
   }
 
-  if(contains(files, ['index.html', '!Dockerfile', '!package.json'])) {
+  if (contains(files, ["index.html", "!Dockerfile", "!package.json"])) {
     return {
-      type: 'static',
+      type: "static",
       build: `
         FROM mhart/alpine-node:base-8
         WORKDIR ${directory}
         ADD . .
 
         CMD ["node", "index.js"]
-      `
+      `,
     };
   }
 
   return {
-    type: 'unknown'
+    type: "unknown",
   };
 }
 
-module.exports = classifier;
+export default classifier;
