@@ -1,16 +1,14 @@
-const tar = require('tar');
-const path = require('path');
-const fs = require('fs');
-
-const { promisify } = require('util');
+import tar from "tar";
+import path from "path";
+import fs from "fs";
+import { promisify } from "util";
 
 const writeFile = promisify(fs.writeFile);
 const readFile = promisify(fs.readFile);
 
-const classifer = require('./classifier');
-
-const { getPort, mk } = require('./helpers/util');
-const { get, build, remove } = require('./models/deployment');
+import classifer from "./classifier.js";
+import { getPort, mk } from "./helpers/util.js";
+import Deployment from "./models/deployment.js";
 
 /**
  * handles the deployment of an application tar
@@ -21,8 +19,13 @@ const { get, build, remove } = require('./models/deployment');
  * @param {String} option.token      - token associated with the user that want to deploy the application
  * @param {String} option.username   - username of the user the token is associated too
  */
-module.exports = async function deploy({ name, bundlePath, token, username }) {
-  const deployment = await get({ username, token, name, create: true });
+export default async function deploy({ name, bundlePath, token, username }) {
+  const deployment = await Deployment.get({
+    username,
+    token,
+    name,
+    create: true,
+  });
   const outputDir = `${process.cwd()}/tmp/${deployment.subdomain}`;
 
   // makes the directory recursively
@@ -30,33 +33,38 @@ module.exports = async function deploy({ name, bundlePath, token, username }) {
 
   await tar.x({
     file: bundlePath,
-    cwd: outputDir
+    cwd: outputDir,
   });
 
-  await remove({
+  await Deployment.remove({
     token,
     username,
-    subdomain: deployment.subdomain
+    subdomain: deployment.subdomain,
   });
 
   const config = await classifer(outputDir);
 
-  if (config.type === 'unknown') {
-    throw new Error('deployment not supported');
+  if (config.type === "unknown") {
+    throw new Error("deployment not supported");
   }
 
-  if (config.type === 'static') {
-    await writeFile(path.resolve(outputDir, 'index.js'), (await readFile(path.resolve(__dirname, 'static', 'static-server.js'))).toString('utf8'));
+  if (config.type === "static") {
+    await writeFile(
+      path.resolve(outputDir, "index.js"),
+      (
+        await readFile(new URL("static", "static-server.js", import.meta.url))
+      ).toString("utf8")
+    );
   }
 
-  await writeFile(path.resolve(outputDir, 'Dockerfile'), config.build);
+  await writeFile(path.resolve(outputDir, "Dockerfile"), config.build);
 
-  return await build({
+  return await Deployment.build({
     name,
     token,
     username,
     subdomain: deployment.subdomain,
     port: await getPort(),
-    directory: outputDir
+    directory: outputDir,
   });
-};
+}
