@@ -1,10 +1,11 @@
-const tar = require('tar');
-const os = require('os');
-const fs = require('fs');
-const path = require('path');
-const { promisify } = require('util');
+import tar from "tar";
+import os from "os";
+import fs from "fs";
+import path from "path";
+import { promisify } from "util";
+import FormData from "form-data";
 
-const { request } = require('./util');
+import { request } from "./util.js";
 
 const unlink = promisify(fs.unlink);
 const writeFile = promisify(fs.writeFile);
@@ -32,11 +33,14 @@ class CLI {
    * @return {Promise}
    */
   createBundle(directory) {
-    return tar.c({
-      gzip: true,
-      portable: true,
-      file: 'bundle.tgz'
-    }, fs.readdirSync(directory));
+    return tar.c(
+      {
+        gzip: true,
+        portable: true,
+        file: "bundle.tgz",
+      },
+      fs.readdirSync(directory)
+    );
   }
   /**
    * removes the bundle from the given directory
@@ -46,7 +50,7 @@ class CLI {
    * @return {Promise}
    */
   async removeBundle(directory) {
-    return await unlink(path.resolve(directory, 'bundle.tgz'));
+    return await unlink(path.resolve(directory, "bundle.tgz"));
   }
   /**
    * Deals with uploading a specified bundle
@@ -58,16 +62,19 @@ class CLI {
    * @return {Promise}
    */
   async uploadBundle({ name, bundle, token, username }) {
-    return await request('post', {
+    const form = new FormData();
+
+    form.append("name", name);
+    form.append("bundle", bundle);
+
+    return await request("post", {
       url: `${this.url}/upload`,
       headers: {
-        'x-deploy-token': token,
-        'x-deploy-username': username
+        "x-deploy-token": token,
+        "x-deploy-username": username,
+        ...form.getHeaders(),
       },
-      formData: {
-        name,
-        bundle
-      }
+      data: form.getBuffer(),
     });
   }
   /**
@@ -80,12 +87,12 @@ class CLI {
    * @return {Promise}
    */
   async login({ username, password }) {
-    return await request('post', {
+    return await request("post", {
       url: `${this.url}/login`,
-      form: {
+      data: {
         username,
-        password
-      }
+        password,
+      },
     });
   }
   /**
@@ -98,12 +105,12 @@ class CLI {
    * @return {Promise}
    */
   async register({ username, password }) {
-    return await request('post', {
+    return await request("post", {
       url: `${this.url}/register`,
-      form: {
+      data: {
         username,
-        password
-      }
+        password,
+      },
     });
   }
   /**
@@ -116,12 +123,12 @@ class CLI {
    * @return {Promise}
    */
   async logout({ token, username }) {
-    return await request('get', {
+    return await request("get", {
       url: `${this.url}/api/logout`,
       headers: {
-        'x-deploy-token': token,
-        'x-deploy-username': username
-      }
+        "x-deploy-token": token,
+        "x-deploy-username": username,
+      },
     });
   }
   /**
@@ -135,12 +142,12 @@ class CLI {
    * @return {Promise}
    */
   async getLogs({ token, username, name }) {
-    return await request('get', {
+    return await request("get", {
       url: `${this.url}/api/deployments/${name}/logs`,
       headers: {
-        'x-deploy-token': token,
-        'x-deploy-username': username
-      }
+        "x-deploy-token": token,
+        "x-deploy-username": username,
+      },
     });
   }
   /**
@@ -155,14 +162,14 @@ class CLI {
    */
   async getDeployments({ token, username, name }) {
     let uri = `${this.url}/api/deployments`;
-    if(name) uri += `/${name}`;
+    if (name) uri += `/${name}`;
 
-    return await request('get', {
+    return await request("get", {
       url: uri,
       headers: {
-        'x-deploy-token': token,
-        'x-deploy-username': username
-      }
+        "x-deploy-token": token,
+        "x-deploy-username": username,
+      },
     });
   }
   /**
@@ -176,12 +183,12 @@ class CLI {
    * @return {Promise}
    */
   async deleteDeployment({ name, token, username }) {
-    return await request('delete', {
+    return await request("delete", {
       url: `${this.url}/api/deployments/${name}`,
       headers: {
-        'x-deploy-token': token,
-        'x-deploy-username': username
-      }
+        "x-deploy-token": token,
+        "x-deploy-username": username,
+      },
     });
   }
   /**
@@ -193,12 +200,12 @@ class CLI {
    * @return {Promise}
    */
   async getUserDetails({ token, username }) {
-    return await request('get', {
+    return await request("get", {
       url: `${this.url}/api/user`,
       headers: {
-        'x-deploy-token': token,
-        'x-deploy-username': username
-      }
+        "x-deploy-token": token,
+        "x-deploy-username": username,
+      },
     });
   }
   /**
@@ -213,15 +220,15 @@ class CLI {
   async cacheCredentials({ username, token }) {
     const credentials = {
       username,
-      token
+      token,
     };
-    const configPath = path.resolve(os.homedir(), '.deployrc');
+    const configPath = path.resolve(os.homedir(), ".deployrc");
 
     await writeFile(configPath, JSON.stringify(credentials, null, 4));
 
     return {
       username,
-      token
+      token,
     };
   }
   /**
@@ -232,24 +239,24 @@ class CLI {
    */
   async getCredentials() {
     try {
-      const configPath = path.resolve(os.homedir(), '.deployrc');
+      const configPath = path.resolve(os.homedir(), ".deployrc");
       const credentials = JSON.parse(await readFile(configPath));
       const { username, token } = credentials;
-      if(username && token) {
+      if (username && token) {
         return {
           username,
-          token
+          token,
         };
       } else {
-        throw new Error('credentials not found');
+        throw new Error("credentials not found");
       }
-    } catch(ex) {
-      if(ex.code === 'ENOENT') {
-        throw new Error('You are not logged in, please login to view identity');
+    } catch (ex) {
+      if (ex.code === "ENOENT") {
+        throw new Error("You are not logged in, please login to view identity");
       }
       throw new Error(ex);
     }
   }
 }
 
-module.exports = CLI;
+export default CLI;
