@@ -1,10 +1,11 @@
 /**
  * @module lib/helpers/util
  */
-const net = require('net');
-const fs = require('fs');
-const { promisify } = require('util');
-const _request = require('request');
+import net from "net";
+
+import fs from "fs";
+import { promisify } from "util";
+import axios from "axios";
 
 const readdir = promisify(fs.readdir);
 const stat = promisify(fs.stat);
@@ -19,20 +20,20 @@ let portrange = 45032;
  * @method getPort
  * @return {Promise}
  */
-module.exports.getPort = function getPort() {
+export const getPort = function getPort() {
   return new Promise((resolve, reject) => {
     try {
       portrange += 1;
       var server = net.createServer();
 
       server.listen(portrange, () => {
-        server.once('close', () => {
+        server.once("close", () => {
           return resolve(portrange);
         });
         server.close();
       });
-      server.on('error', () => resolve(getPort()));
-    } catch(ex) {
+      server.on("error", () => resolve(getPort()));
+    } catch (ex) {
       return reject(ex);
     }
   });
@@ -43,23 +44,23 @@ module.exports.getPort = function getPort() {
  * @param  {String} directory - path to future directory
  * @return {Promise}
  */
-module.exports.mk = async function mk(directory) {
-  let cwd = '';
-  let path = directory.split('/');
+export const mk = async function mk(directory) {
+  let cwd = "";
+  let path = directory.split("/");
   // If the first value is blank that means we were given an absolute path and should respect that
-  if(path[0] == '') {
+  if (path[0] == "") {
     path[1] = `/${path[1]}`;
   }
   // remove all of the empty entries
-  path = path.filter((p) => p !== '');
+  path = path.filter((p) => p !== "");
 
   for (var i = 0; i <= path.length - 1; i++) {
-    cwd += `${path[i]}${i === path.length - 1 ? '' : '/'}`;
+    cwd += `${path[i]}${i === path.length - 1 ? "" : "/"}`;
     try {
       // try to open the directory, if we can't we will create it
       await readdir(cwd);
-    } catch(ex) {
-      if(ex.message.indexOf('no such file or directory') > -1) {
+    } catch (ex) {
+      if (ex.message.indexOf("no such file or directory") > -1) {
         await mkdir(cwd);
       }
     }
@@ -75,7 +76,7 @@ module.exports.mk = async function mk(directory) {
  */
 const possible = "abcdefghijklmnopqrstuvwxyz0123456789";
 
-module.exports.hash = function hash(length) {
+export const hash = function hash(length) {
   let text = "";
 
   for (let i = 0; i < length; i++) {
@@ -85,15 +86,13 @@ module.exports.hash = function hash(length) {
   return text;
 };
 
-module.exports.request = function request(type, options) {
-  return new Promise(function(resolve, reject) {
-    _request[type](options, (err, httpResponse, body) => {
-      if(err) reject(err);
-      const response = JSON.parse(body);
-      if(response.error) return reject(response.error);
-      return resolve(response);
-    });
+export const request = async function request(type, options) {
+  const { data } = await axios({
+    method: type,
+    ...options,
   });
+
+  return data;
 };
 
 /**
@@ -105,11 +104,18 @@ module.exports.request = function request(type, options) {
  * @example
  * contains(['index.html', 'main.css'], ['index.html', '!Dockerfile', '!package.json'])
  */
-module.exports.contains = function contains(arr, contains) {
+export const contains = function contains(arr, contains) {
   var conditions = [];
   for (var i in contains) {
-    var key = contains[i].substring(0, 1) === '!' ? contains[i].substring(1, contains[i].length) : contains[i];
-    conditions.push(contains[i].substring(0, 1) === '!' ? arr.indexOf(key) === -1 : arr.indexOf(key) > -1);
+    var key =
+      contains[i].substring(0, 1) === "!"
+        ? contains[i].substring(1, contains[i].length)
+        : contains[i];
+    conditions.push(
+      contains[i].substring(0, 1) === "!"
+        ? arr.indexOf(key) === -1
+        : arr.indexOf(key) > -1
+    );
   }
   if (conditions.indexOf(false) > -1) {
     return false;
@@ -117,12 +123,12 @@ module.exports.contains = function contains(arr, contains) {
   return true;
 };
 
-module.exports.rm = async function rm(dir) {
+export const rm = async function rm(dir) {
   let files = await readdir(dir);
 
   if (files.length > 0) {
     for (var i = 0; i < files.length; i++) {
-      var file = dir + '/' + files[i];
+      var file = dir + "/" + files[i];
       if ((await stat(file)).isFile()) {
         await unlink(file);
       } else {
