@@ -1,36 +1,38 @@
-const mongoose = require('mongoose');
-const http = require('http');
-const fs = require('fs');
-const path = require('path');
-const Docker = require('dockerode');
+import mongoose from "mongoose";
+import http from "http";
+import fs from "fs";
+import Docker from "dockerode";
 const docker = new Docker({
-  socketPath: '/var/run/docker.sock'
+  socketPath: "/var/run/docker.sock",
 });
-const stream = require('stream');
+import stream from "stream";
 
 const Schema = mongoose.Schema;
 
-const { authenticate } = require('./user');
-const { hash, rm } = require('../helpers/util');
-const { count } = require('./request');
-const deleteRequests = require('./request').del;
+import User from "./user.js";
+import Request from "./request.js";
 
-const DeploymentSchema = new Schema({
-  id: String,
-  name: String,
-  port: Number,
-  subdomain: String,
-  directory: String,
-  username: String
-}, {
-  minimize: false,
-  timestamps: {
-    createdAt: 'created_at',
-    updatedAt: 'updated_at'
+import { hash, rm } from "../helpers/util.js";
+
+const DeploymentSchema = new Schema(
+  {
+    id: String,
+    name: String,
+    port: Number,
+    subdomain: String,
+    directory: String,
+    username: String,
+  },
+  {
+    minimize: false,
+    timestamps: {
+      createdAt: "created_at",
+      updatedAt: "updated_at",
+    },
   }
-});
+);
 
-const DeploymentModel = mongoose.model('Deployment', DeploymentSchema);
+const DeploymentModel = mongoose.model("Deployment", DeploymentSchema);
 
 /**
  * Deployment definition
@@ -53,15 +55,14 @@ class Deployment {
    * @return {Promise}
    */
   static async update({ name, token, username, deployment }) {
-    let user = await authenticate(username, token);
+    let user = await User.authenticate(username, token);
 
     let found = await DeploymentModel.findOne({
       username: user.username,
-      name
+      name,
     });
 
-    if(found) {
-
+    if (found) {
       found.id = deployment.id || found.id;
       found.name = deployment.name || found.name;
       found.port = deployment.port || found.port;
@@ -76,7 +77,7 @@ class Deployment {
       const deployment = {
         name,
         subdomain: `${name}-${hash(6)}`,
-        username: user.username
+        username: user.username,
       };
       return await DeploymentModel.create(deployment);
     }
@@ -91,11 +92,11 @@ class Deployment {
    * @return {Promise}
    */
   static async del({ name, username, token }) {
-    const user = await authenticate(username, token);
+    const user = await User.authenticate(username, token);
 
     const deployment = await DeploymentModel.findOne({
       username: user.username,
-      name
+      name,
     });
 
     await rm(deployment.directory);
@@ -103,13 +104,13 @@ class Deployment {
     await Deployment.remove({
       subdomain: deployment.subdomain,
       token,
-      username: user.username
+      username: user.username,
     });
 
-    await deployment.remove();
+    await deployment.deleteOne();
 
-    await deleteRequests({
-      subdomain: deployment.subdomain
+    await Request.del({
+      subdomain: deployment.subdomain,
     });
   }
   /**
@@ -124,16 +125,18 @@ class Deployment {
     const { host } = headers;
 
     // If this is not an upload request, it is a proxy request
-    const subdomain = host.split('.')[0];
+    const subdomain = host.split(".")[0];
 
     let deployment = await DeploymentModel.findOne({
-      "subdomain":{
-        "$eq": subdomain
-      }
+      subdomain: {
+        $eq: subdomain,
+      },
     });
 
-    if(!deployment) {
-      return res.status(404).sendFile(path.resolve(__dirname, '../static/not-found.html'));
+    if (!deployment) {
+      return res
+        .status(404)
+        .sendFile(new URL("../static/not-found.html", import.meta.url));
     }
 
     const { port } = deployment;
@@ -142,18 +145,22 @@ class Deployment {
       path: url,
       headers,
       port,
-      host: 'localhost'
+      host: "localhost",
     });
-    proxy.addListener('response', function (proxy_response) {
-      proxy_response.addListener('data', (chunk) => res.write(chunk, 'binary'));
-      proxy_response.addListener('end', () => res.end());
+    proxy.addListener("response", function (proxy_response) {
+      proxy_response.addListener("data", (chunk) => res.write(chunk, "binary"));
+      proxy_response.addListener("end", () => res.end());
       res.writeHead(proxy_response.statusCode, proxy_response.headers);
     });
-    proxy.on('error', () => {
-      res.status(502).sendFile(path.resolve(__dirname, '../static/page-could-not-load.html'));
+    proxy.on("error", () => {
+      res
+        .status(502)
+        .sendFile(
+          new URL("../static/page-could-not-load.html", import.meta.url)
+        );
     });
-    req.addListener('data', (chunk) => proxy.write(chunk, 'binary'));
-    req.addListener('end', () => proxy.end());
+    req.addListener("data", (chunk) => proxy.write(chunk, "binary"));
+    req.addListener("end", () => proxy.end());
   }
   /**
    * decorates a deployment with the correct data on get
@@ -163,7 +170,7 @@ class Deployment {
    * @return {Promise}
    */
   static async decorate(deployment) {
-    let requests = await count({ subdomain: deployment.subdomain });
+    let requests = await Request.count({ subdomain: deployment.subdomain });
 
     deployment.requests = requests || 0;
 
@@ -173,8 +180,8 @@ class Deployment {
       let info = await container.inspect();
 
       deployment.status = info.State.Status;
-    } catch(ex) {
-      deployment.status = 'error';
+    } catch (ex) {
+      deployment.status = "error";
     }
 
     return deployment;
@@ -191,18 +198,18 @@ class Deployment {
    * @return {Promise}
    */
   static async get({ name, token, username, create }) {
-    let user = await authenticate(username, token);
+    let user = await User.authenticate(username, token);
 
     let deployment = await DeploymentModel.findOne({
       username: user.username,
-      name
+      name,
     });
 
-    if(deployment) {
+    if (deployment) {
       return await Deployment.decorate(deployment._doc);
     }
 
-    if(create) {
+    if (create) {
       // add a new deployment name
       // TODO: figure out how to avoid collisions
       return await DeploymentModel.create({
@@ -221,13 +228,13 @@ class Deployment {
    * @return {Promise}
    */
   static async getAll({ token, username }) {
-    let user = await authenticate(username, token);
+    let user = await User.authenticate(username, token);
 
     let deployments = await DeploymentModel.find({
-      username: user.username
+      username: user.username,
     });
 
-    for(var i = 0; i < deployments.length; i++) {
+    for (var i = 0; i < deployments.length; i++) {
       deployments[i] = await Deployment.decorate(deployments[i]._doc);
     }
 
@@ -238,7 +245,7 @@ class Deployment {
       let files = [];
       fs.readdirSync(directory).map((f) => {
         let stat = fs.statSync(`${directory}/${f}`);
-        if(stat.isDirectory()) {
+        if (stat.isDirectory()) {
           files = files.concat(recurse(`${directory}/${f}`));
         } else {
           files.push(`${directory}/${f}`);
@@ -247,61 +254,68 @@ class Deployment {
       return files;
     }
     // we are working off the the relative paths and we need to make the absolute paths relative to context
-    let src = recurse(directory).map((f) => f.replace(directory, ''));
+    let src = recurse(directory).map((f) => f.replace(directory, ""));
 
-    return new Promise(function(resolve, reject) {
-      docker.buildImage({
-        context: directory,
-        src
-      }, {
-        t: subdomain
-      }, (err, stream) => {
-        if(err) return reject(err);
-        docker.modem.followProgress(stream, onFinished, onProgress);
+    return new Promise(function (resolve, reject) {
+      docker.buildImage(
+        {
+          context: directory,
+          src,
+        },
+        {
+          t: subdomain,
+        },
+        (err, stream) => {
+          if (err) return reject(err);
+          docker.modem.followProgress(stream, onFinished, onProgress);
 
-        function onFinished(err) {
-          if(err) return reject(err);
-          docker.createContainer({
-            Image: subdomain,
-            name: subdomain,
-            env: [
-              'PORT=3000'
-            ],
-            ExposedPorts: {
-              '3000/tcp': {}
-            },
-            PortBindings: {
-              '3000/tcp': [{
-                'HostPort': `${port}`
-              }]
-            },
-            Privileged: true
-          }, (err, container) => {
+          function onFinished(err) {
             if (err) return reject(err);
-            container.start((err) => {
-              if (err) return reject(err);
-              const id = container.id;
+            docker.createContainer(
+              {
+                Image: subdomain,
+                name: subdomain,
+                env: ["PORT=3000"],
+                ExposedPorts: {
+                  "3000/tcp": {},
+                },
+                PortBindings: {
+                  "3000/tcp": [
+                    {
+                      HostPort: `${port}`,
+                    },
+                  ],
+                },
+                Privileged: true,
+              },
+              (err, container) => {
+                if (err) return reject(err);
+                container.start((err) => {
+                  if (err) return reject(err);
+                  const id = container.id;
 
-              Deployment.update({
-                name,
-                username,
-                token,
-                deployment: {
-                  id,
-                  port,
-                  directory
-                }
-              })
-                .then((deployment) => resolve(deployment))
-                .catch((ex) => reject(ex));
-            });
-          });
+                  Deployment.update({
+                    name,
+                    username,
+                    token,
+                    deployment: {
+                      id,
+                      port,
+                      directory,
+                    },
+                  })
+                    .then((deployment) => resolve(deployment))
+                    .catch((ex) => reject(ex));
+                });
+              }
+            );
+          }
+          // TODO: be able to stream the output of this to a socket to give real time updates
+          function onProgress() {
+            // console.log(ev);
+          }
         }
-        // TODO: be able to stream the output of this to a socket to give real time updates
-        function onProgress() {
-          // console.log(ev);
-        }
-      });
+      );
     });
   }
   /**
@@ -316,21 +330,21 @@ class Deployment {
   static async start({ name, username, token }) {
     let opts = {};
     // TODO: check username and token to make sure the request is authenticated
-    if(username && token) {
+    if (username && token) {
       opts.username = username;
     }
 
     const deployments = await DeploymentModel.find(opts);
 
-    if(deployments.length === 0) return;
+    if (deployments.length === 0) return;
 
-    for(var i = 0; i < deployments.length; i++) {
+    for (var i = 0; i < deployments.length; i++) {
       const deployment = deployments[i];
 
-      if(name && deployment.name == name) {
+      if (name && deployment.name == name) {
         await docker.getContainer(deployment.id).start();
       }
-      if(!name) {
+      if (!name) {
         await docker.getContainer(deployment.id).start();
       }
     }
@@ -348,21 +362,21 @@ class Deployment {
   static async stop({ name, token, username }) {
     let opts = {};
     // TODO: check username and token to make sure the request is authenticated
-    if(username && token) {
+    if (username && token) {
       opts.username = username;
     }
 
     const deployments = await DeploymentModel.find(opts);
 
-    if(deployments.length === 0) return;
+    if (deployments.length === 0) return;
 
-    for(var i = 0; i < deployments.length; i++) {
+    for (var i = 0; i < deployments.length; i++) {
       const deployment = deployments[i];
 
-      if(name && deployment.name == name) {
+      if (name && deployment.name == name) {
         await docker.getContainer(deployment.id).kill({ force: true });
       }
-      if(!name) {
+      if (!name) {
         await docker.getContainer(deployment.id).kill({ force: true });
       }
     }
@@ -379,13 +393,14 @@ class Deployment {
    * @return {Promise}
    */
   static logs({ name, token, username }) {
+    // eslint-disable-next-line no-async-promise-executor
     return new Promise(async (resolve, reject) => {
       try {
-        const user = await authenticate(username, token);
+        const user = await User.authenticate(username, token);
 
         const deployment = await DeploymentModel.findOne({
           name,
-          username: user.username
+          username: user.username,
         });
 
         const logs = [];
@@ -393,23 +408,27 @@ class Deployment {
 
         const logStream = new stream.PassThrough();
 
-        logStream.on('data', (chunk) => { logs.push(chunk.toString('utf8')); });
+        logStream.on("data", (chunk) => {
+          logs.push(chunk.toString("utf8"));
+        });
 
         const cStream = await container.logs({
           follow: true,
           stdout: true,
-          stderr: true
+          stderr: true,
         });
 
         container.modem.demuxStream(cStream, logStream, logStream);
-        cStream.on('end', function(){
+        cStream.on("end", function () {
           return resolve(logs);
         });
 
-        setTimeout(function() {
+        setTimeout(function () {
           cStream.destroy();
+
+          return resolve(logs);
         }, 2000);
-      } catch(ex) {
+      } catch (ex) {
         return reject(ex);
       }
     });
@@ -424,11 +443,11 @@ class Deployment {
    * @return {Promise}
    */
   static async remove({ subdomain, token, username }) {
-    let user = await authenticate(username, token);
+    let user = await User.authenticate(username, token);
 
     let deployment = await DeploymentModel.findOne({
       subdomain: subdomain,
-      username: user.username
+      username: user.username,
     });
 
     try {
@@ -439,11 +458,11 @@ class Deployment {
       let info = await container.inspect();
 
       await docker.getImage(info.Image).remove({
-        force: true
+        force: true,
       });
       await container.remove();
-    } catch(ex) {
-      if(ex.reason !== 'no such container') {
+    } catch (ex) {
+      if (ex.reason !== "no such container") {
         throw new Error(ex);
       }
     }
@@ -452,4 +471,4 @@ class Deployment {
   }
 }
 
-module.exports = Deployment;
+export default Deployment;
