@@ -1,25 +1,28 @@
-const mongoose = require('mongoose');
+import mongoose from "mongoose";
 
 const Schema = mongoose.Schema;
 
-const RequestSchema = new Schema({
-  url: String,
-  time: Number,
-  subdomain: String,
-  statusCode: String,
-  userAgent: String,
-  referer: String,
-  method: String,
-  acceptLanguage: String
-}, {
-  minimize: false,
-  timestamps: {
-    createdAt: 'created_at',
-    updatedAt: 'updated_at'
+const RequestSchema = new Schema(
+  {
+    url: String,
+    time: Number,
+    subdomain: String,
+    statusCode: String,
+    userAgent: String,
+    referer: String,
+    method: String,
+    acceptLanguage: String,
+  },
+  {
+    minimize: false,
+    timestamps: {
+      createdAt: "created_at",
+      updatedAt: "updated_at",
+    },
   }
-});
+);
 
-const RequestModel = mongoose.model('Request', RequestSchema);
+const RequestModel = mongoose.model("Request", RequestSchema);
 
 /**
  * Request definition
@@ -46,10 +49,10 @@ class Request {
     const { url, headers, method } = req;
     const { host } = headers;
 
-    const subdomain = host.split('.')[0];
+    const subdomain = host.split(".")[0];
     const startAt = process.hrtime();
 
-    res.on('finish', () => {
+    res.on("finish", () => {
       const diff = process.hrtime(startAt);
       const time = diff[0] * 1e3 + diff[1] * 1e-6;
 
@@ -61,9 +64,9 @@ class Request {
         subdomain,
         method,
         statusCode,
-        userAgent: headers['user-agent'],
-        referer: headers['referer'],
-        acceptLanguage: headers['accept-language']
+        userAgent: headers["user-agent"],
+        referer: headers["referer"],
+        acceptLanguage: headers["accept-language"],
       });
     });
 
@@ -78,7 +81,7 @@ class Request {
    */
   static async count({ subdomain }) {
     return await RequestModel.count({
-      subdomain
+      subdomain,
     });
   }
   /**
@@ -89,10 +92,10 @@ class Request {
    * @return {Promise}
    */
   static async del({ subdomain }) {
-    return await RequestModel.remove({
-      subdomain
+    return await RequestModel.deleteMany({
+      subdomain,
     });
   }
 }
 
-module.exports = Request;
+export default Request;
