@@ -1,22 +1,25 @@
-const mongoose = require('mongoose');
-const crypto = require('crypto');
+import mongoose from "mongoose";
+import crypto from "crypto";
 
 const Schema = mongoose.Schema;
 
-const UserSchema = new Schema({
-  _id: String,
-  username: String,
-  password: String,
-  token: String
-}, {
-  minimize: false,
-  timestamps: {
-    createdAt: 'created_at',
-    updatedAt: 'updated_at'
+const UserSchema = new Schema(
+  {
+    _id: String,
+    username: String,
+    password: String,
+    token: String,
+  },
+  {
+    minimize: false,
+    timestamps: {
+      createdAt: "created_at",
+      updatedAt: "updated_at",
+    },
   }
-});
+);
 
-const UserModel = mongoose.model('User', UserSchema);
+const UserModel = mongoose.model("User", UserSchema);
 
 /**
  * User definition
@@ -35,11 +38,11 @@ class User {
    * @param {Function} next - callback to go to next middleware
    */
   static async authenticateMiddleware(req, res, next) {
-    const username = req.headers['x-deploy-username'];
-    const token = req.headers['x-deploy-token'];
+    const username = req.headers["x-deploy-username"];
+    const token = req.headers["x-deploy-token"];
 
-    if(!username && !token) {
-      res.status(500).send({ error: 'authentication necessary' });
+    if (!username && !token) {
+      res.status(500).send({ error: "authentication necessary" });
     }
 
     req.user = await User.authenticate(username, token);
@@ -57,14 +60,14 @@ class User {
   static async authenticate(username, token) {
     let user = await UserModel.findOne({
       _id: username,
-      token
+      token,
     });
 
-    if(!user) throw new Error('not authenticated');
+    if (!user) throw new Error("not authenticated");
 
     return {
       username: user.username,
-      token: user.token
+      token: user.token,
     };
   }
   /**
@@ -78,12 +81,12 @@ class User {
   static async logout({ token, username }) {
     let user = await UserModel.findOne({
       _id: username,
-      token
+      token,
     });
 
-    if(!user) throw new Error('token and username not valid');
+    if (!user) throw new Error("token and username not valid");
 
-    user.token = '';
+    user.token = "";
 
     return await user.save();
   }
@@ -98,17 +101,20 @@ class User {
   static async login({ username, password }) {
     let user = await UserModel.findOne({
       _id: username,
-      password: crypto.createHash('sha256').update(password).digest('hex')
+      password: crypto.createHash("sha256").update(password).digest("hex"),
     });
 
-    if(!user) throw new Error('password username combination not correct');
+    if (!user) throw new Error("password username combination not correct");
 
-    user.token = crypto.createHash('sha256').update(Date.now() + password + username).digest('hex');
+    user.token = crypto
+      .createHash("sha256")
+      .update(Date.now() + password + username)
+      .digest("hex");
     await user.save();
 
     return {
       username: user.username,
-      token: user.token
+      token: user.token,
     };
   }
   /**
@@ -123,16 +129,19 @@ class User {
     let user = await UserModel.create({
       _id: username,
       username,
-      password: crypto.createHash('sha256').update(password).digest('hex'),
-      token: crypto.createHash('sha256').update(Date.now() + password + username).digest('hex')
+      password: crypto.createHash("sha256").update(password).digest("hex"),
+      token: crypto
+        .createHash("sha256")
+        .update(Date.now() + password + username)
+        .digest("hex"),
     });
 
     return {
       username: user.username,
       token: user.token,
-      depoyments: user.deployments
+      depoyments: user.deployments,
     };
   }
 }
 
-module.exports = User;
+export default User;
