@@ -1,13 +1,13 @@
-const express = require('express');
-const formidable = require('formidable');
-const bodyParser = require('body-parser');
+import express from "express";
+import formidable from "formidable";
+import bodyParser from "body-parser";
 
 const app = express();
 
-const deploy = require('./deploy');
-const { authenticateMiddleware, register, logout, login } = require('./models/user');
-const { del, logs, get, getAll, proxy } = require('./models/deployment');
-const { log } = require('./models/request');
+import deploy from "./deploy.js";
+import User from "./models/user.js";
+import Deployment from "./models/deployment.js";
+import Request from "./models/request.js";
 
 const port = process.env.PORT || 5000;
 
@@ -30,135 +30,181 @@ const formidablePromise = (req, opts) => {
   });
 };
 
-app.post('/register', asyncMiddleware(async (req, res) => {
-  const { username, password } = req.body;
-  if(!username || !password) return res.status(500).send({ error: 'please provide username and password'});
+app.post(
+  "/register",
+  asyncMiddleware(async (req, res) => {
+    const { username, password } = req.body;
 
-  try {
-    const user = await register({ username, password });
-    res.status(200).send(user);
-  } catch(error) {
-    res.status(500).send({ error: 'something went wrong' });
-  }
-}));
+    console.log(username, password);
+    if (!username || !password)
+      return res
+        .status(500)
+        .send({ error: "please provide username and password" });
 
-app.post('/login', asyncMiddleware(async (req, res) => {
-  const { username, password } = req.body;
+    try {
+      const user = await User.register({ username, password });
+      res.status(200).send(user);
+    } catch (error) {
+      console.log(error);
+      res.status(500).send({ error: "something went wrong" });
+    }
+  })
+);
 
-  if(!username || !password) return res.status(500).send({ error: 'please provide username and password'});
+app.post(
+  "/login",
+  asyncMiddleware(async (req, res) => {
+    const { username, password } = req.body;
 
-  try {
-    const user = await login({ username, password });
-    res.status(200).send(user);
-  } catch(error) {
-    res.status(500).send({ error });
-  }
-}));
+    if (!username || !password)
+      return res
+        .status(500)
+        .send({ error: "please provide username and password" });
 
-app.post('/upload', authenticateMiddleware, asyncMiddleware(async (req, res) => {
-  try {
-    const { user } = req;
-    const { username, token } = user;
+    try {
+      const user = await User.login({ username, password });
+      res.status(200).send(user);
+    } catch (error) {
+      res.status(500).send({ error });
+    }
+  })
+);
 
-    const { fields, files } = await formidablePromise(req, {
-      type: true
-    });
+app.post(
+  "/upload",
+  User.authenticateMiddleware,
+  asyncMiddleware(async (req, res) => {
+    try {
+      const { user } = req;
+      const { username, token } = user;
 
-    const { name } = fields;
-    const { bundle } = files;
-    const deployment = await deploy({
-      name,
-      bundlePath: bundle.path,
-      token,
-      username
-    });
-    res.status(200).send({ success: `application ${name} deployed successfully`, deployment });
-  } catch(ex) {
-    res.status(500).send({ error: ex.stack });
-  }
-}));
+      const { fields, files } = await formidablePromise(req, {
+        type: true,
+      });
 
-app.get('/api/deployments/:name/logs', authenticateMiddleware, asyncMiddleware(async (req, res) => {
-  const { name } = req.params;
-  const { token, username } = req.user;
+      const { name } = fields;
+      const { bundle } = files;
 
-  try {
-    const _logs = await logs({
-      name,
-      username,
-      token
-    });
-    res.status(200).send({ logs: _logs });
-  } catch(ex) {
-    res.status(500).send({ ex });
-  }
-}));
+      const deployment = await deploy({
+        name,
+        bundlePath: bundle.filepath,
+        token,
+        username,
+      });
+      res.status(200).send({
+        success: `application ${name} deployed successfully`,
+        deployment,
+      });
+    } catch (ex) {
+      console.log(ex);
+      res.status(500).send({ error: ex.stack });
+    }
+  })
+);
 
-app.get('/api/deployments', authenticateMiddleware, asyncMiddleware(async (req, res) => {
-  const { token, username } = req.user;
+app.get(
+  "/api/deployments/:name/logs",
+  User.authenticateMiddleware,
+  asyncMiddleware(async (req, res) => {
+    const { name } = req.params;
+    const { token, username } = req.user;
 
-  try {
-    const deployments = await getAll({
-      username,
-      token
-    });
-    res.status(200).send({ deployments });
-  } catch(error) {
-    res.status(500).send({ error });
-  }
-}));
+    try {
+      const logs = await Deployment.logs({
+        name,
+        username,
+        token,
+      });
+      res.status(200).send({ logs });
+    } catch (ex) {
+      res.status(500).send({ ex });
+    }
+  })
+);
 
-app.delete('/api/deployments/:name', authenticateMiddleware, asyncMiddleware(async (req, res) => {
-  const { name } = req.params;
-  const { token, username } = req.user;
+app.get(
+  "/api/deployments",
+  User.authenticateMiddleware,
+  asyncMiddleware(async (req, res) => {
+    const { token, username } = req.user;
 
-  try {
-    await del({
-      name,
-      token,
-      username
-    });
-    res.status(200).send({ success: `deployment ${name} successfully deleted` });
-  } catch(error) {
-    res.status(500).send({ error });
-  }
-}));
+    try {
+      const deployments = await Deployment.getAll({
+        username,
+        token,
+      });
+      res.status(200).send({ deployments });
+    } catch (error) {
+      res.status(500).send({ error });
+    }
+  })
+);
 
-app.get('/api/deployments/:name', authenticateMiddleware, asyncMiddleware(async (req, res) => {
-  const { name } = req.params;
-  const { token, username } = req.user;
+app.delete(
+  "/api/deployments/:name",
+  User.authenticateMiddleware,
+  asyncMiddleware(async (req, res) => {
+    const { name } = req.params;
+    const { token, username } = req.user;
 
-  try {
-    const deployment = await get({
-      username,
-      token,
-      name
-    });
-    res.status(200).send({ deployment });
-  } catch(error) {
-    res.status(500).send({ error });
-  }
-}));
+    try {
+      await Deployment.del({
+        name,
+        token,
+        username,
+      });
+      res
+        .status(200)
+        .send({ success: `deployment ${name} successfully deleted` });
+    } catch (error) {
+      res.status(500).send({ error });
+    }
+  })
+);
 
-app.get('/api/logout', authenticateMiddleware, asyncMiddleware(async (req, res) => {
-  const { token, username } = req.user;
+app.get(
+  "/api/deployments/:name",
+  User.authenticateMiddleware,
+  asyncMiddleware(async (req, res) => {
+    const { name } = req.params;
+    const { token, username } = req.user;
 
-  try {
-    await logout({ username, token });
-    res.status(200).send({ success: `${username} successfully deleted` });
-  } catch(ex) {
-    res.status(500).send({ error: 'could not logout from session' });
-  }
-}));
+    try {
+      const deployment = await Deployment.get({
+        username,
+        token,
+        name,
+      });
+      res.status(200).send({ deployment });
+    } catch (error) {
+      res.status(500).send({ error });
+    }
+  })
+);
 
-app.get('/api/user', authenticateMiddleware, (req, res) => {
+app.get(
+  "/api/logout",
+  User.authenticateMiddleware,
+  asyncMiddleware(async (req, res) => {
+    const { token, username } = req.user;
+
+    try {
+      await User.logout({ username, token });
+      res.status(200).send({ success: `${username} successfully deleted` });
+    } catch (ex) {
+      res.status(500).send({ error: "could not logout from session" });
+    }
+  })
+);
+
+app.get("/api/user", User.authenticateMiddleware, (req, res) => {
   const { user } = req;
 
-  delete user['token'];
+  delete user["token"];
   res.status(200).send({ user });
 });
 
-app.get('*', log, proxy);
+app.get("*", Request.log, Deployment.proxy);
 
 app.listen(port, () => {
   console.log(`⛅️ deploy.sh is running on port ${port}`); // eslint-disable-line
