@@ -1,18 +1,20 @@
 #!/usr/bin/env node
+import moment from "moment";
+import Url from "url";
+import Table from "turtler";
+import ora from "ora";
 
-const moment = require('moment');
-const Url = require('url');
-const Table = require('turtler');
+export default async function (cli) {
+  const spinner = ora().start();
 
-module.exports = async function(cli, spinner) {
-  spinner.text = 'Getting deployment list';
+  spinner.text = "Getting deployment list";
 
   const { token, username } = await cli.getCredentials();
   const { deployments } = await cli.getDeployments({ token, username });
 
   spinner.stop();
 
-  if(deployments.length > 0) {
+  if (deployments.length > 0) {
     let data = [["name", "url", "age", "requests", "status"]];
 
     deployments.forEach((r, i) => {
@@ -20,13 +22,19 @@ module.exports = async function(cli, spinner) {
       config.host = `${r.subdomain}.${config.host}`;
       const url = Url.format(config);
 
-      data[i + 1] = [r.name, url, moment(r.updated_at).fromNow(), r.requests.toString(), r.status];
+      data[i + 1] = [
+        r.name,
+        url,
+        moment(r.updated_at).fromNow(),
+        r.requests.toString(),
+        r.status,
+      ];
     });
 
     let table = new Table(data);
 
     console.log(table.markdown()); // eslint-disable-line
   } else {
-    console.log('0 deployments found'); // eslint-disable-line
+    console.log("0 deployments found"); // eslint-disable-line
   }
-};
+}
